requirements.txt
scikit-learn>=1.2.0
jupyterlab>=3.6.0
notebook>=6.5.0
ipykernel>=6.22.0
matplotlib>=3.7.0
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.0
pandas>=1.3.0
numpy<2.0
python-docx>=0.8.0
requests>=2.25.0
python-dotenv>=0.19.0
gradio>=4.0.0